requirements.txt
biopython==1.76 ; python_version <= '2.7' or (python_version >= '3.0' and python_version <= '3.5')
biopython>=1.79 ; (python_version > '2.7' and python_version < '3.0') or python_version > '3.5'
matplotlib>=3.5.0
networkx>=2.6.3
numpy>=1.22.3
python-dateutil>=2.8.2
wget>=3.2
